+++ src/wifi/test/channel-access-manager-test.cc
@@ -22,9 +22,11 @@
 #include "ns3/frame-exchange-manager.h"
 #include "ns3/interference-helper.h"
 #include "ns3/multi-model-spectrum-channel.h"
+#include "ns3/pointer.h"
 #include "ns3/qos-txop.h"
 #include "ns3/simulator.h"
 #include "ns3/spectrum-wifi-phy.h"
+#include "ns3/string.h"
 #include "ns3/test.h"
 
 #include <list>
@@ -638,7 +640,9 @@ ChannelAccessManagerTest<TxopType>::AddTxop(uint32_t aifsn)
     m_txop.push_back(txop);
     m_ChannelAccessManager->Add(txop);
     // the following causes the creation of a link for the txop object
-    auto mac = CreateObject<AdhocWifiMac>();
+    auto mac = CreateObjectWithAttributes<AdhocWifiMac>(
+        "Txop",
+        PointerValue(CreateObjectWithAttributes<Txop>("AcIndex", StringValue("AC_BE_NQOS"))));
     mac->SetWifiPhys({nullptr});
     txop->SetWifiMac(mac);
     txop->SetAifsn(aifsn);