+++ RELEASE_NOTES.md
@@ -30,6 +30,7 @@ been tested on Linux. The latest known version to work with ns-3 is cppyy==3.1.2
 - (lr-wpan) !1924 - Adds MAC attribute identifiers
 - (lr-wpan) !1927 - Adds standard version comments to MLME-GET.request function
 - (lr-wpan) !1926 - Adds namespace lrwpan and prefix shortening
+- (wifi) - The `WifiMacHelper` provides a `SetDcf` and a `SetEdca` methods to configure attributes of `Txop` and `QosTxop` objects, respectively
 
 ### Bugs fixed